feat: add MapClient.tsx, implement dynamic into map/page.tsx
